First (untested) iteration of paper scraper with text-to-speech

+++ run.sh
@@ -13,13 +13,14 @@ source ./env/bin/activate
 python ./src/main.py
 
 mv ./workdir/related_papers.tex "$OUTPUT_FOLDER"
+mv ./workdir/tts_papers.mp3"$OUTPUT_FOLDER"
 
 # Compile the document
 cd "$OUTPUT_FOLDER" || exit 1
 pdflatex -interaction=nonstopmode "related_papers.tex"
 # Cleanup the folder - only leave .tex and .pdf files
 for file in *; do
-    if [[ ! $file =~ \.tex$ ]] && [[ ! $file =~ \.pdf$ ]]; then
+    if [[ ! $file =~ \.tex$ ]] && [[ ! $file =~ \.pdf$ ]] && [[ ! $file =~ \.mp3$ ]]; then
         rm "$file"
     fi
 done
